requirements.txt
tensorflow==2.18.0
numpy==2.0.2
flask==3.1.0
flask-cors==5.0.0
pillow==11.0.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1
google-cloud-storage
gunicorn